build nextjs without errors
