+++ nextclient/src/lib/game-client/DataModel.ts
@@ -10,72 +10,7 @@ export enum Color {
  * A 8x8 board square address in Algebraic notation
  * @see {@link https://en.wikipedia.org/wiki/Algebraic_notation_(chess)}
  */
-export enum Square {
-    A8 = "a8",
-    B8 = "b8",
-    C8 = "c8",
-    D8 = "d8",
-    E8 = "e8",
-    F8 = "f8",
-    G8 = "g8",
-    H8 = "h8",
-    A7 = "a7",
-    B7 = "b7",
-    C7 = "c7",
-    D7 = "d7",
-    E7 = "e7",
-    F7 = "f7",
-    G7 = "g7",
-    H7 = "h7",
-    A6 = "a6",
-    B6 = "b6",
-    C6 = "c6",
-    D6 = "d6",
-    E6 = "e6",
-    F6 = "f6",
-    G6 = "g6",
-    H6 = "h6",
-    A5 = "a5",
-    B5 = "b5",
-    C5 = "c5",
-    D5 = "d5",
-    E5 = "e5",
-    F5 = "f5",
-    G5 = "g5",
-    H5 = "h5",
-    A4 = "a4",
-    B4 = "b4",
-    C4 = "c4",
-    D4 = "d4",
-    E4 = "e4",
-    F4 = "f4",
-    G4 = "g4",
-    H4 = "h4",
-    A3 = "a3",
-    B3 = "b3",
-    C3 = "c3",
-    D3 = "d3",
-    E3 = "e3",
-    F3 = "f3",
-    G3 = "g3",
-    H3 = "h3",
-    A2 = "a2",
-    B2 = "b2",
-    C2 = "c2",
-    D2 = "d2",
-    E2 = "e2",
-    F2 = "f2",
-    G2 = "g2",
-    H2 = "h2",
-    A1 = "a1",
-    B1 = "b1",
-    C1 = "c1",
-    D1 = "d1",
-    E1 = "e1",
-    F1 = "f1",
-    G1 = "g1",
-    H1 = "h1",
-}
+export type Square = "a8" | "b8" | "c8" | "d8" | "e8" | "f8" | "g8" | "h8" | "a7" | "b7" | "c7" | "d7" | "e7" | "f7" | "g7" | "h7" | "a6" | "b6" | "c6" | "d6" | "e6" | "f6" | "g6" | "h6" | "a5" | "b5" | "c5" | "d5" | "e5" | "f5" | "g5" | "h5" | "a4" | "b4" | "c4" | "d4" | "e4" | "f4" | "g4" | "h4" | "a3" | "b3" | "c3" | "d3" | "e3" | "f3" | "g3" | "h3" | "a2" | "b2" | "c2" | "d2" | "e2" | "f2" | "g2" | "h2" | "a1" | "b1" | "c1" | "d1" | "e1" | "f1" | "g1" | "h1";
 
 /**
  * Piece symbol (without color)
